fn_uptime_data_generator/requirements.txt
cognite-sdk == 7.*
pandas >= 2.0.0
pyyaml >= 6.0.1
